# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Support Connect - Live WhatsApp Support Chat App
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Article {
  id: number;
  title: string;
  summary: string;
  /** @nullable */
  content?: string | null;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  publishedAt: string;
  isFeatured: boolean;
  isTrending: boolean;
  viewCount: number;
}

export interface BreakingNews {
  text: string;
}

export interface ChatSession {
  id: string;
  userName: string;
  /** @nullable */
  userPhone?: string | null;
  status: string;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  lastMessage?: string | null;
  unreadCount: number;
}

export interface ChatSessionInput {
  userName: string;
  userPhone?: string;
}

export interface ChatMessage {
  id: number;
  sessionId: string;
  content: string;
  sender: string;
  createdAt: string;
  isRead: boolean;
}

export interface ChatMessageInput {
  content: string;
}

export interface AdminLoginInput {
  username: string;
  password: string;
}

export interface AdminUser {
  id: number;
  username: string;
}

export interface AdminLoginResult {
  success: boolean;
  token: string;
  admin: AdminUser;
}

export interface AdminStats {
  totalSessions: number;
  openSessions: number;
  closedSessions: number;
  totalMessages: number;
  todaySessions: number;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface ErrorResponse {
  error: string;
}

export type ListArticlesParams = {
category?: string;
limit?: number;
};