# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Support Connect - Live WhatsApp Support Chat App
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/me:
    get:
      operationId: getAdminMe
      tags: [admin]
      summary: Get current admin info
      responses:
        "200":
          description: Admin info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/sessions/{sessionId}:
    delete:
      operationId: deleteAdminSession
      tags: [admin]
      summary: Delete a chat session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/sessions/{sessionId}/close:
    post:
      operationId: closeAdminSession
      tags: [admin]
      summary: Close a chat session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSession"
  /admin/sessions/{sessionId}/messages:
    get:
      operationId: listAdminSessionMessages
      tags: [admin]
      summary: Get messages for a session (admin)
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendAdminMessage
      tags: [admin]
      summary: Send admin reply to a session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
...
components:
  schemas:
...
    ChatSession:
      type: object
      properties:
        id:          { type: string }
        userName:    { type: string }
        userPhone:   { type: ["string", "null"] }
        status:      { type: string }
        createdAt:   { type: string }
        updatedAt:   { type: string }
        lastMessage: { type: ["string", "null"] }
        unreadCount: { type: number }
      required: [id, userName, status, createdAt, updatedAt, unreadCount]
...
    ChatMessage:
      type: object
      properties:
        id:        { type: number }
        sessionId: { type: string }
        content:   { type: string }
        sender:    { type: string }
        createdAt: { type: string }
        isRead:    { type: boolean }
      required: [id, sessionId, content, sender, createdAt, isRead]
    ChatMessageInput:
      type: object
      properties:
        content: { type: string }
      required: [content]
    AdminLoginInput:
      type: object
      properties:
        username: { type: string }
        password: { type: string }
      required: [username, password]
    AdminLoginResult:
      type: object
      properties:
        success: { type: boolean }
        token:   { type: string }
        admin:
          $ref: "#/components/schemas/AdminUser"
      required: [success, token, admin]
    AdminUser:
      type: object
      properties:
        id:       { type: number }
        username: { type: string }
      required: [id, username]
...
    SuccessResponse:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
      required: [success]
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]