# lib/api-zod/src/generated/types/adminLoginInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Support Connect - Live WhatsApp Support Chat App
 * OpenAPI spec version: 0.1.0
 */

export interface AdminLoginInput {
  username: string;
  password: string;
}