# lib/api-zod/src/generated/types/adminLoginResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Support Connect - Live WhatsApp Support Chat App
 * OpenAPI spec version: 0.1.0
 */
import type { AdminUser } from './adminUser';

export interface AdminLoginResult {
  success: boolean;
  token: string;
  admin: AdminUser;
}