# lib/api-zod/src/generated/types/chatMessage.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Support Connect - Live WhatsApp Support Chat App
 * OpenAPI spec version: 0.1.0
 */

export interface ChatMessage {
  id: number;
  sessionId: string;
  content: string;
  sender: string;
  createdAt: string;
  isRead: boolean;
}