# lib/api-zod/src/generated/types/chatSession.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Support Connect - Live WhatsApp Support Chat App
 * OpenAPI spec version: 0.1.0
 */

export interface ChatSession {
  id: string;
  userName: string;
  /** @nullable */
  userPhone?: string | null;
  status: string;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  lastMessage?: string | null;
  unreadCount: number;
}